+++ src/main/java/org/thinkingstudio/mafglib/util/NeoUtils.java
@@ -1,0 +1,27 @@
+package org.thinkingstudio.mafglib.util;
+
+import net.minecraft.client.gui.screen.Screen;
+import net.neoforged.fml.IExtensionPoint;
+import net.neoforged.fml.ModContainer;
+import net.neoforged.neoforge.client.ConfigScreenHandler;
+
+import java.util.function.Function;
+
+public class NeoUtils {
+    private static NeoUtils INSTANCE;
+
+    public void getClientModIgnoredServerOnly(ModContainer modContainer) {
+        modContainer.registerExtensionPoint(IExtensionPoint.DisplayTest.class, () -> new IExtensionPoint.DisplayTest(() -> IExtensionPoint.DisplayTest.IGNORESERVERONLY, (a, b) -> true));
+    }
+
+    public void registerModConfigScreen(ModContainer modContainer, Function<Screen, Screen> screenFunction) {
+        modContainer.registerExtensionPoint(ConfigScreenHandler.ConfigScreenFactory.class, () -> new ConfigScreenHandler.ConfigScreenFactory((client, screen) -> screenFunction.apply(screen)));
+    }
+
+    public static NeoUtils getInstance() {
+        if (INSTANCE == null) {
+            INSTANCE = new NeoUtils();
+        }
+        return INSTANCE;
+    }
+}